requirements.txt
google.generativeai
uuid
pydantic
fastapi
requests
streamlit
google.generativeai
uuid
pydantic
uvicorn
